Updated names and organized directories.
I made the directory structure make significantly more sense. For instance i had 4 asset directories which i consolodated into a single directory with subdirectories for different asset categories. This will make it easier to follow and easier to modify for any later developers who work on this project as well as for myself working forward. This will also make it easier to document changes to the theme and to other tools as the project progresses without much overhead on the part of any team members or the project sponsor.